fix sinking ship problem and resetting misses
when getting a new set of hits, don't reset the misses if the ship
types are equal. this fixes the problem of a shooter forgetting
about previous misses. if the shiptype is new, do a full reset.

when a ship was sunk, it was based off of shiptype. the shiptype
that the ship was targeting was a best guess and is probably
different than an actual sinking ship. determine if a ship was
sunk based off of actual location, not ship type